Implement cleanData function
We want cleanData to return a lean object that is flexible enough for
whatever client use case the spreadsheet is being streamed to.

cleanData now an object whose keys are names of sheets and values are
arrays of the sheet’s rows. Each row is an object whose values are
keyed to the sheet’s column names.

This commit should not have taken eight hours, but it took me eight
hours. One day we will laugh about this.